fix:Login&Registerページのデザイン
