+++ src/app/(page)/(dashboard)/layout.tsx
@@ -11,11 +11,11 @@ export default function MainLayout({
       <div className="md:hidden">
         <Header />
       </div>
-      <main className="flex bg-gray-100 min-h-screen">
+      <main className="flex bg-gray-100 min-h-screen standalone-top-adjust">
         <div className="fixed hidden md:block">
           <Navigation />
         </div>
-        <div className="relative w-full md:ml-64 top-16 md:top-0 standalone-top-adjust">
+        <div className="relative w-full md:ml-64 top-16 md:top-0">
           {children}
         </div>
         <div className="standalone-only">
