--- src/app/(page)/login/_components/LoginForm.tsx
@@ -1,138 +1,0 @@
-"use client";
-
-import { useState, FormEvent } from "react";
-import { useRouter, useSearchParams } from "next/navigation";
-import Link from "next/link";
-import { useAuth } from "@/context/AuthContext";
-import { ROUTES } from "@/constants";
-
-export default function LoginForm() {
-  const router = useRouter();
-  const searchParams = useSearchParams();
-  const { login } = useAuth();
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-
-  // クエリパラメータから登録完了メッセージを取得
-  const registeredMessage = searchParams.get("registered")
-    ? "アカウント登録が完了しました。ログインしてください。"
-    : null;
-
-  const [formData, setFormData] = useState({
-    email: "",
-    password: "",
-  });
-
-  const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const { name, value } = e.target;
-    setFormData((prev) => ({ ...prev, [name]: value }));
-  };
-
-  const handleSubmit = async (e: FormEvent) => {
-    e.preventDefault();
-    setError(null);
-
-    // 入力検証
-    if (!formData.email || !formData.password) {
-      setError("メールアドレスとパスワードを入力してください");
-      return;
-    }
-
-    try {
-      setLoading(true);
-      const result = await login(formData.email, formData.password);
-
-      if (result.success) {
-        // ログイン成功時、マイページにリダイレクト
-        router.push(ROUTES.MYPAGE);
-      } else {
-        setError(result.error || "ログインに失敗しました");
-      }
-    } catch (err) {
-      setError("ログイン処理中にエラーが発生しました");
-      console.error("ログインエラー:", err);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  return (
-    <form className="space-y-6" onSubmit={handleSubmit}>
-      {error && (
-        <div className="bg-red-50 border-l-4 border-red-500 p-4">
-          <p className="text-sm text-red-700">{error}</p>
-        </div>
-      )}
-
-      {registeredMessage && (
-        <div className="bg-green-50 border-l-4 border-green-500 p-4">
-          <p className="text-sm text-green-700">{registeredMessage}</p>
-        </div>
-      )}
-
-      <div>
-        <label
-          htmlFor="email"
-          className="block text-sm font-medium text-gray-700"
-        >
-          メールアドレス
-        </label>
-        <div className="mt-1">
-          <input
-            id="email"
-            name="email"
-            type="email"
-            autoComplete="email"
-            required
-            value={formData.email}
-            onChange={handleChange}
-            className="appearance-none block w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm placeholder-gray-400 focus:outline-none focus:ring-blue-500 focus:border-blue-500 sm:text-sm"
-          />
-        </div>
-      </div>
-
-      <div>
-        <label
-          htmlFor="password"
-          className="block text-sm font-medium text-gray-700"
-        >
-          パスワード
-        </label>
-        <div className="mt-1">
-          <input
-            id="password"
-            name="password"
-            type="password"
-            autoComplete="current-password"
-            required
-            value={formData.password}
-            onChange={handleChange}
-            className="appearance-none block w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm placeholder-gray-400 focus:outline-none focus:ring-blue-500 focus:border-blue-500 sm:text-sm"
-          />
-        </div>
-      </div>
-
-      <div>
-        <button
-          type="submit"
-          disabled={loading}
-          className="w-full flex justify-center py-2 px-4 border border-transparent rounded-md shadow-sm text-sm font-medium text-white bg-blue-600 hover:bg-blue-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500 disabled:opacity-50"
-        >
-          {loading ? "処理中..." : "ログイン"}
-        </button>
-      </div>
-
-      <div className="text-sm text-center mt-4">
-        <p className="text-gray-600">
-          アカウントをお持ちでないですか？{" "}
-          <Link
-            href="/register"
-            className="font-medium text-blue-600 hover:text-blue-500"
-          >
-            アカウント登録
-          </Link>
-        </p>
-      </div>
-    </form>
-  );
-}
