--- src/app/(page)/login/page.tsx
@@ -1,28 +1,0 @@
-import { Metadata } from "next";
-import LoginForm from "./_components/LoginForm";
-import { Suspense } from "react";
-
-export const metadata: Metadata = {
-  title: "ログイン | アプリ名",
-  description: "アカウントにログインしてサービスをご利用ください。",
-};
-
-export default function LoginPage() {
-  return (
-    <div className="min-h-screen bg-gray-50 flex flex-col justify-center py-12 sm:px-6 lg:px-8">
-      <div className="sm:mx-auto sm:w-full sm:max-w-md">
-        <h1 className="text-center text-3xl font-extrabold text-gray-900">
-          ログイン
-        </h1>
-      </div>
-
-      <div className="mt-8 sm:mx-auto sm:w-full sm:max-w-md">
-        <div className="bg-white py-8 px-4 shadow sm:rounded-lg sm:px-10">
-          <Suspense fallback={<div>Loading...</div>}>
-            <LoginForm />
-          </Suspense>
-        </div>
-      </div>
-    </div>
-  );
-}
