--- src/app/(page)/register/_components/RegisterForm.tsx
@@ -1,183 +1,0 @@
-"use client";
-
-import { useState, FormEvent } from "react";
-import { useRouter } from "next/navigation";
-import Link from "next/link";
-import { useAuth } from "@/context/AuthContext";
-import { ROUTES } from "@/constants";
-
-export default function RegisterForm() {
-  const router = useRouter();
-  const { register } = useAuth();
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-
-  const [formData, setFormData] = useState({
-    userName: "",
-    email: "",
-    password: "",
-    confirmPassword: "",
-  });
-
-  const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const { name, value } = e.target;
-    setFormData((prev) => ({ ...prev, [name]: value }));
-  };
-
-  const handleSubmit = async (e: FormEvent) => {
-    e.preventDefault();
-    setError(null);
-
-    // 入力検証
-    if (!formData.userName || !formData.email || !formData.password) {
-      setError("すべての項目を入力してください");
-      return;
-    }
-
-    if (formData.password !== formData.confirmPassword) {
-      setError("パスワードが一致しません");
-      return;
-    }
-
-    if (formData.password.length < 6) {
-      setError("パスワードは6文字以上で入力してください");
-      return;
-    }
-
-    try {
-      setLoading(true);
-      const result = await register(
-        formData.email,
-        formData.password,
-        formData.userName
-      );
-
-      if (result.success) {
-        // 登録成功時、マイページにリダイレクト
-        router.push(ROUTES.MYPAGE);
-      } else {
-        setError(result.error || "登録に失敗しました");
-      }
-    } catch (err) {
-      setError("登録処理中にエラーが発生しました");
-      console.error("登録エラー:", err);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  return (
-    <form className="space-y-6" onSubmit={handleSubmit}>
-      {error && (
-        <div className="bg-red-50 border-l-4 border-red-500 p-4">
-          <p className="text-sm text-red-700">{error}</p>
-        </div>
-      )}
-
-      <div>
-        <label
-          htmlFor="userName"
-          className="block text-sm font-medium text-gray-700"
-        >
-          ユーザー名
-        </label>
-        <div className="mt-1">
-          <input
-            id="userName"
-            name="userName"
-            type="text"
-            required
-            value={formData.userName}
-            onChange={handleChange}
-            className="appearance-none block w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm placeholder-gray-400 focus:outline-none focus:ring-blue-500 focus:border-blue-500 sm:text-sm"
-          />
-        </div>
-      </div>
-
-      <div>
-        <label
-          htmlFor="email"
-          className="block text-sm font-medium text-gray-700"
-        >
-          メールアドレス
-        </label>
-        <div className="mt-1">
-          <input
-            id="email"
-            name="email"
-            type="email"
-            autoComplete="email"
-            required
-            value={formData.email}
-            onChange={handleChange}
-            className="appearance-none block w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm placeholder-gray-400 focus:outline-none focus:ring-blue-500 focus:border-blue-500 sm:text-sm"
-          />
-        </div>
-      </div>
-
-      <div>
-        <label
-          htmlFor="password"
-          className="block text-sm font-medium text-gray-700"
-        >
-          パスワード
-        </label>
-        <div className="mt-1">
-          <input
-            id="password"
-            name="password"
-            type="password"
-            autoComplete="new-password"
-            required
-            value={formData.password}
-            onChange={handleChange}
-            className="appearance-none block w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm placeholder-gray-400 focus:outline-none focus:ring-blue-500 focus:border-blue-500 sm:text-sm"
-          />
-        </div>
-      </div>
-
-      <div>
-        <label
-          htmlFor="confirmPassword"
-          className="block text-sm font-medium text-gray-700"
-        >
-          パスワード（確認）
-        </label>
-        <div className="mt-1">
-          <input
-            id="confirmPassword"
-            name="confirmPassword"
-            type="password"
-            autoComplete="new-password"
-            required
-            value={formData.confirmPassword}
-            onChange={handleChange}
-            className="appearance-none block w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm placeholder-gray-400 focus:outline-none focus:ring-blue-500 focus:border-blue-500 sm:text-sm"
-          />
-        </div>
-      </div>
-
-      <div>
-        <button
-          type="submit"
-          disabled={loading}
-          className="w-full flex justify-center py-2 px-4 border border-transparent rounded-md shadow-sm text-sm font-medium text-white bg-blue-600 hover:bg-blue-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500 disabled:opacity-50"
-        >
-          {loading ? "処理中..." : "登録する"}
-        </button>
-      </div>
-
-      <div className="text-sm text-center mt-4">
-        <p className="text-gray-600">
-          すでにアカウントをお持ちですか？{" "}
-          <Link
-            href={ROUTES.LOGIN}
-            className="font-medium text-blue-600 hover:text-blue-500"
-          >
-            ログイン
-          </Link>
-        </p>
-      </div>
-    </form>
-  );
-}
