--- src/app/(page)/register/page.tsx
@@ -1,25 +1,0 @@
-import { Metadata } from "next";
-import RegisterForm from "./_components/RegisterForm";
-
-export const metadata: Metadata = {
-  title: "新規登録 | Score Graph",
-  description: "アカウントを作成して、サービスをご利用ください。",
-};
-
-export default function RegisterPage() {
-  return (
-    <div className="min-h-screen bg-gray-50 flex flex-col justify-center py-12 sm:px-6 lg:px-8">
-      <div className="sm:mx-auto sm:w-full sm:max-w-md">
-        <h1 className="text-center text-3xl font-extrabold text-gray-900">
-          アカウント作成
-        </h1>
-      </div>
-
-      <div className="mt-8 sm:mx-auto sm:w-full sm:max-w-md">
-        <div className="bg-white py-8 px-4 shadow sm:rounded-lg sm:px-10">
-          <RegisterForm />
-        </div>
-      </div>
-    </div>
-  );
-}
