+++ src/components/general/BottomNavigation.tsx
@@ -11,7 +11,7 @@ export default function BottomNavigation() {
   const { user, loading } = useAuth();
 
   return (
-    <div className="fixed bottom-0 left-0 z-11 w-full h-20 bg-white border-t border-gray-200 px-4 pb-8">
+    <div className="fixed bottom-0 left-0 z-11 w-full h-20 bg-white border-t border-gray-200 px-4 pb-8 text-gray-700">
       <div className="grid h-full grid-cols-3 mx-auto max-w-xl">
         <button
           onClick={() => {
@@ -21,7 +21,7 @@ export default function BottomNavigation() {
         >
           <HomeIcon
             className={`w-7 h-7 ${
-              pathname === ROUTES.HOME ? "font-bold" : "font-normal"
+              pathname === ROUTES.HOME ? "font-bold" : ""
             }`}
             strokeWidth={pathname === ROUTES.HOME ? 2.5 : 1.5}
           />
@@ -34,7 +34,7 @@ export default function BottomNavigation() {
         >
           <Newspaper
             className={`w-7 h-7 ${
-              pathname === ROUTES.NEWS ? "font-bold" : "font-normal"
+              pathname === ROUTES.NEWS ? "font-bold" : ""
             }`}
             strokeWidth={pathname === ROUTES.NEWS ? 2.5 : 1.5}
           />
