+++ src/constants/index.ts
@@ -1,17 +1,16 @@
-// import { Time } from '@/libs/time'
-
 export const ROUTES = {
-  TOP: '/',
-  HOME: '/home',
-  HOME_TESTREAD: 'home/testRead',
-  FRIEND: '/friend',
-  NEWS: '/news',
-  MYPAGE: '/mypage',
-  PROFILE_UPDATE: '/mypage/profileUpdate',
-  CARD_CREATE: 'mypage/cardCreate',
-  CARD_UPDATE: 'mypage/cardUpdate',
-  MYPAGE_TESTREAD: 'mypage/testRead',
-  TEST_CREATE: 'mypage/testCreate',
-  TARGET_UPSERT: 'mypage/targetUpsert',
-  LOGIN: '/login',
-} as const
+  TOP: "/",
+  HOME: "/home",
+  HOME_TESTREAD: "home/testRead",
+  FRIEND: "/friend",
+  NEWS: "/news",
+  MYPAGE: "/mypage",
+  PROFILE_UPDATE: "/mypage/profileUpdate",
+  CARD_CREATE: "mypage/cardCreate",
+  CARD_UPDATE: "mypage/cardUpdate",
+  MYPAGE_TESTREAD: "mypage/testRead",
+  TEST_CREATE: "mypage/testCreate",
+  TARGET_UPSERT: "mypage/targetUpsert",
+  LOGIN: "/login",
+  REGISTER: "/register",
+} as const;
